+++ lib/chat/__tests__/getChatUrl.test.ts
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+import { getChatUrl } from "@/lib/chat/getChatUrl";
+
+describe("getChatUrl", () => {
+  it("builds an absolute session-scoped chat URL", () => {
+    expect(getChatUrl("sess-1", "chat-2")).toBe(
+      "https://chat.recoupable.com/sessions/sess-1/chats/chat-2",
+    );
+  });
+});
