+++ lib/chat/getChatPath.ts
@@ -1,0 +1,3 @@
+export function getChatPath(sessionId: string, chatId: string): string {
+  return `/sessions/${sessionId}/chats/${chatId}`;
+}
