+++ lib/chat/getChatUrl.ts
@@ -1,0 +1,5 @@
+import { getChatPath } from "@/lib/chat/getChatPath";
+
+export function getChatUrl(sessionId: string, chatId: string): string {
+  return `https://chat.recoupable.com${getChatPath(sessionId, chatId)}`;
+}
